+++ tests/basic/relaxed_destruction_order.cpp
@@ -5,8 +5,9 @@
 #include <iostream>
 
 struct LogProvider : public wis::LogLayer {
-    virtual void Log(wis::Severity sev, std::string message, wis::source_location sl = wis::source_location::current()) override {
-        // std::cout << wis::format("[{}]: {}\n", wis::severity_strings[+sev], message);
+    virtual void Log(wis::Severity sev, std::string message, wis::source_location sl = wis::source_location::current()) override
+    {
+        std::cout << wis::format("[{}]: {}\n", wis::severity_strings[+sev], message);
     };
 };
 
@@ -119,5 +120,41 @@ TEST_CASE("basic_device")
         a = {};
         REQUIRE_FALSE(error);
         REQUIRE_FALSE(a);
+    }
+
+    SECTION("texture replacement")
+    {
+        wis::Texture a;
+        auto [res6, texture] = allocator.CreateTexture(wis::TextureDesc{ .format = wis::DataFormat::RGBA8Unorm,
+                                                                         .size = {
+
+                                                                                 .width = 1024,
+                                                                                 .height = 1024,
+                                                                                 .depth_or_layers = 1,
+                                                                         } });
+        REQUIRE(texture);
+
+        a = std::move(texture);
+        REQUIRE(a);
+        REQUIRE_FALSE(texture);
+
+        auto [res7, texture2] = allocator.CreateTexture(wis::TextureDesc{ .format = wis::DataFormat::RGBA8Unorm,
+                                                                          .size = {
+
+                                                                                  .width = 1024,
+                                                                                  .height = 1024,
+                                                                                  .depth_or_layers = 1,
+                                                                          } });
+
+        REQUIRE(texture2);
+        texture = std::move(texture2);
+        REQUIRE(texture);
+        REQUIRE_FALSE(texture2);
+
+        a = std::move(texture);
+        REQUIRE(a);
+        REQUIRE_FALSE(texture);
+
+
     }
 }